split up logic into multiple files
