+++ sympy/logic/algorithms/forward_chaining_theory.py
@@ -8,416 +8,13 @@ from types import MappingProxyType
 from functools import reduce
 from sympy.assumptions.cnf import Literal
 
+from sympy.logic.algorithms.rules_engine import RulesEngine
+from sympy.assumptions.facts2 import (id_rules_dict, rules_dict, id_direct_dict,
+                                      id_to_pred, direct_dict, pred_to_id, pred_id_to_bitvec,
+                                      pred_id_neg_to_direct_implicants_bitset)
 
-from sympy.core.symbol import Symbol
-from sympy.strategies.core import switch
-from collections.abc import Iterable
 
-
-class BitSet:
-    def __init__(self, value=0):
-        """Initialize the BitSet with an integer value (default: 0)."""
-        self.value = value
-
-    def add(self, x):
-        """Add an element to the bit set."""
-        self.value |= (1 << x)
-
-    def remove(self, x):
-        """Remove an element from the bit set."""
-        self.value &= ~(1 << x)
-
-    def contains(self, x):
-        """Check if an element is in the bit set."""
-        return (self.value & (1 << x)) != 0
-
-    def toggle(self, x):
-        """Toggle (flip) an element in the bit set."""
-        self.value ^= (1 << x)
-
-    def size(self):
-        """Return the number of elements in the bit set."""
-        return bin(self.value).count('1')
-
-    def to_list(self):
-        """Convert the bit set to a list of integers."""
-        return [i for i in range(self.value.bit_length()) if self.contains(i)]
-
-    def clear(self):
-        """Clear the bit set (reset to empty)."""
-        self.value = 0
-
-    def union(self, other):
-        """Return a new BitSet with the union of two sets."""
-        return BitSet(self.value | other.value)
-
-    def intersection(self, other):
-        """Return a new BitSet with the intersection of two sets."""
-        return BitSet(self.value & other.value)
-
-    def difference(self, other):
-        """Return a new BitSet with the difference (self - other)."""
-        return BitSet(self.value & ~other.value)
-
-    def __repr__(self):
-        """String representation of the BitSet."""
-        return f"BitSet({self.to_list()})"
-
-def _AppliedPredicate_to_Predicate(pred):
-    if isinstance(pred, AppliedPredicate):
-        return pred.function
-    if isinstance(pred, Predicate):
-        return pred
-    if isinstance(pred, BooleanFunction):
-        return pred.func(*[_AppliedPredicate_to_Predicate(arg) for arg in pred.args])
-
-    raise ValueError(f"Object {pred} is not of a recognized type")
-
-def _is_literal(expr):
-    return isinstance(expr, Predicate) or (isinstance(expr.args[0], Predicate) and isinstance(expr, Not)), f"{expr} is not a literal"
-
-def _add_rule(rules_dict, antecedent, implicant, remove_var = True):
-
-
-
-    if remove_var:
-        antecedent = _AppliedPredicate_to_Predicate(antecedent)
-        implicant = _AppliedPredicate_to_Predicate(implicant)
-    if isinstance(antecedent, Predicate) and isinstance(implicant, Predicate):
-        rules_dict[(antecedent,)].add(implicant)
-        rules_dict[(to_nnf(~implicant),)].add(to_nnf(~antecedent))  # contrapositive
-        if len([key for key in rules_dict.keys() if not isinstance(key, Iterable)]) > 0:
-            assert len([key for key in rules_dict.keys() if not isinstance(key, Iterable)]) == 0
-        return
-
-    antecedent = to_nnf(antecedent, simplify=False)
-    implicant = to_nnf(implicant, simplify=False)
-    if isinstance(antecedent, Or) or isinstance(implicant, And):
-        if isinstance(implicant, And):
-            antecedent, implicant = to_nnf(~implicant), to_nnf(~antecedent)
-
-        for disjunct in antecedent.args:
-            _add_rule(rules_dict, disjunct, implicant, remove_var=False)
-
-        if len([key for key in rules_dict.keys() if not isinstance(key, Iterable)]) > 0:
-            assert len([key for key in rules_dict.keys() if not isinstance(key, Iterable)]) == 0
-        return
-
-    assert _is_literal(antecedent) or (all(_is_literal(arg) for arg in antecedent.args) and isinstance(antecedent, And)), antecedent
-
-    if  isinstance(antecedent, And) or isinstance(implicant, Or):
-        if isinstance(implicant, Or):
-            antecedent, implicant = to_nnf(~implicant), to_nnf(~antecedent)
-
-        assert isinstance(antecedent, And)
-        rules_dict[antecedent.args].add(implicant)
-
-        assert len(antecedent.args) >= 2
-        for i in range(len(antecedent.args)):
-            new = And(*antecedent.args[:i] + antecedent.args[i+1:])
-            assert isinstance(And(~implicant, new), And)
-            rules_dict[And(~implicant, new).args].add(~antecedent.args[i])
-        if len([key for key in rules_dict.keys() if not isinstance(key, Iterable)]) > 0:
-            assert len([key for key in rules_dict.keys() if not isinstance(key, Iterable)]) == 0
-
-    else:
-        assert not isinstance(to_nnf(antecedent), And) and not isinstance(to_nnf(antecedent), Or)
-        assert not isinstance(to_nnf(implicant), And) and not isinstance(to_nnf(implicant), Or)
-        rules_dict[(antecedent,)].add(implicant)
-        rules_dict[(to_nnf(~implicant),)].add(~antecedent)
-        if len([key for key in rules_dict.keys() if not isinstance(key, Iterable)]) > 0:
-            assert len([key for key in rules_dict.keys() if not isinstance(key, Iterable)]) == 0
-
-    if len([key for key in rules_dict.keys() if not isinstance(key, Iterable)]) > 0:
-        assert len([key for key in rules_dict.keys() if not isinstance(key, Iterable)]) == 0
-
-from sympy.assumptions import Q
-# add additional rules:
-additional_rules = And(
-    Implies(Q.nonzero, ~Q.zero),
-    Implies(Q.nonpositive, ~Q.positive),
-    Implies(Q.nonnegative, ~Q.negative),
-    Implies(~Q.real, ~Q.nonzero),
-    Implies(~Q.real, ~Q.nonpositive),
-    Implies(~Q.real, ~Q.nonnegative),
-)
-#@cacheit
-def facts_to_dictionary(x = None):
-
-    rules_dict = defaultdict(set)
-
-    facts = And(get_number_facts(), additional_rules)
-
-    for fact in list(facts.args):
-        if isinstance(fact, Implies):
-            _add_rule(rules_dict, fact.args[0], fact.args[1])
-        elif isinstance(fact, Equivalent):
-            _add_rule(rules_dict, fact.args[0], fact.args[1])
-            _add_rule(rules_dict, fact.args[1], fact.args[0])
-        elif isinstance(fact, Not) and isinstance(fact.args[0], And):
-            fact = fact.args[0]
-            _add_rule(rules_dict, fact.args[0], ~fact.args[1])
-        else:
-            raise ValueError(f"{fact} is not of a recognized form")
-
-    composite_predicate = get_composite_predicates()
-    for superset, subsets in composite_predicate.items():
-        _add_rule(rules_dict, subsets, superset)
-        _add_rule(rules_dict, superset, subsets)
-        # for subset in subsets.args:
-        #     _add_rule(rules_dict, subset, superset)
-
-
-
-
-    return rules_dict
-
-
-from collections import deque
-
-
-def transitive_closure(graph):
-    closure = {node: set(neighbors) for node, neighbors in graph.items()}
-
-    for start_node in graph:
-        visited = set()
-        queue = deque([start_node])
-
-        while queue:
-            node = queue.popleft()
-            if node in visited or len(node) > 1:
-                continue
-            visited.add(node)
-
-            for neighbor in graph.get(node, []):
-                if neighbor not in visited:
-                    queue.append((neighbor,))
-                    closure[start_node].add(neighbor)
-
-    return {node: set(neighbors) for node, neighbors in closure.items()}
-
-
-
-from sympy.assumptions.ask_generated import get_known_facts_dict
-
-rules_dict = facts_to_dictionary()
-rules_dict = transitive_closure(rules_dict)
-rules_dict = dict(sorted(rules_dict.items(), key=lambda item: str(item)))
-
-# dictionary of rules with only one antecedent
-# all literals imply themselves
-all_lits = set.union(*rules_dict.values(), {key[0] for key in rules_dict})
-direct_dict = defaultdict(set, {key[0]: val | {key[0]} for key, val in rules_dict.items() if len(key) == 1})
-for lit in all_lits:
-    direct_dict[lit].add(lit)
-
-all_preds = [lit for lit in all_lits if not isinstance(lit, Not)]
-all_preds = sorted(all_preds, key=lambda pred: str(pred))
-
-# if we give more commonly used preds lower numbers, this will help
-pred_to_id = {all_preds[i] : i for i in range(len(all_preds))}
-def pred_to_id_neg_tup(pred):
-    if isinstance(pred, Not):
-        return -pred_to_id[pred.args[0]]#, True
-    else:
-        return pred_to_id[pred]#, False
-
-id_to_pred = [pred for pred, _ in sorted(pred_to_id.items(), key=lambda x: x[1])]
-num_preds = len(all_preds)
-
-id_direct_dict = {pred_to_id_neg_tup(pred) : {pred_to_id_neg_tup(im) for im in imps } for pred, imps in direct_dict.items()}
-id_rules_dict = {}
-for ante, imps in rules_dict.items():
-    id_rules_dict[tuple(pred_to_id_neg_tup(pred) for pred in ante)] = set(pred_to_id_neg_tup(imp) for imp in imps)
-
-#id_rules_dict = { tuple(pred_to_id_neg_tup(pred) for pred in ante) : {pred_to_id_neg_tup(im) for im in imps} for ante, imps in rules_dict.items()}
-
-
-# def pred_lit_set_to_bit_set(pred_lit_set):
-#     i = pred_lit_set_to_int(pred_lit_set)
-#     return BitSet(i)
-
-
-
-def pred_lit_set_to_int(pred_lit_set):
-    binary = []
-    for i in range(2*len(all_preds)):
-        neg = i >= len(all_preds)
-        lit = id_to_pred[i % len(all_preds)]
-        if neg:
-            lit = ~lit
-
-        in_set = (lit in pred_lit_set)*1
-        binary.append(str(in_set))
-
-    return int("".join(binary), 2)
-
-direct_dict_bitset = {}
-for lit in all_lits:
-    direct_imps_bitset = pred_lit_set_to_int(direct_dict[lit])
-    neg = False
-    if isinstance(lit, Not):
-        pred_id = pred_to_id[lit.args[0]]
-        lit = -pred_to_id[lit.args[0]]
-        neg = True
-    else:
-        pred_id = pred_to_id[lit]
-
-    direct_dict_bitset[(pred_id, neg)] = direct_imps_bitset
-
-
-pred_id_to_bitvec = [pred_lit_set_to_int({pred}) for pred in id_to_pred]
-pred_id_to_bitvec += [pred_lit_set_to_int({~pred}) for pred in id_to_pred]
-pred_id_neg_to_direct_implicants_bitset = direct_dict_bitset
-
-#pred_id_direct_dict = {pred_to_id[pred] : 2  for pred, implications in direct_dict.items()}
-
-# I want a mapping of pred number to an int representing the negated preds implied by it
-# I need to give negated preds a seperate number
-
-# I can create an int representing a signle pred or negated pred
-# then I can AND that with the int representing the current state.
-# If non-zero, then the pred is in the set.
-
-direct_dict = dict(direct_dict)
-direct_dict = MappingProxyType(direct_dict)
-
-
-from collections.abc import Iterable
-
-# class RuleTree:
-#     def __init__(self, rules):
-
-
-class RulesEngine:
-    def __init__(self, dictionary):
-        """Initialize the rules engine with a nested dictionary structure."""
-        self.rule_tree = {}
-        self.knowledge_base = {}
-
-        for key, value in dictionary.items():
-            self.add_rule(key, value)
-
-        #self.original_keys = list(self.rule_tree.keys())
-
-        self.rules = self.rule_tree.copy()
-        self.direct = direct_dict
-
-    def reset_state(self):
-        self.rules = self.rule_tree.copy()
-        self.knowledge_base = {}
-
-        #assert list(self.rule_tree.keys()) == self.original_keys
-
-
-    def add_rule(self, conditions, consequence):
-        """
-        Add a rule to the rules engine.
-
-        :param conditions: A set of conditions (e.g., {"a", "b", "c", "d"})
-        :param consequence: The resulting fact (e.g., "e")
-        """
-        rule_tree = self.rule_tree
-        self.print_steps = False
-
-        for condition in sorted(conditions, key=lambda x: str(x)):  # Sort to maintain consistency
-            rule_tree = rule_tree.setdefault(condition, {})
-        rule_tree["__result__"] = (consequence, conditions)   # Store the consequence
-
-    def add_fact(self, fact, source_facts):
-        """Add a fact to the knowledge base."""
-        print(f"adding fact: {self._to_pred(fact)}")
-        self.knowledge_base[fact] = source_facts
-
-    def add_facts(self, facts):
-        """Add a fact to the knowledge base."""
-        if self.print_steps:
-            print(f"adding facts: {[self._to_pred(f)for f in facts]}")
-        self.knowledge_base.update(facts)
-
-    def _to_pred(self, literal):
-        pred_id, neg = literal.arg, literal.is_Not
-        pred = id_to_pred[pred_id]
-        return ~pred if neg else pred
-
-    def trigger(self, trig_fact):
-        if trig_fact not in self.rules or trig_fact not in self.knowledge_base:
-            return False
-
-        implicants, antecedents = [], []
-        pending_facts = set()
-        next = self.rules.pop(trig_fact)
-        for rule_root in next:
-            if rule_root != "__result__":
-                self.rules.update({rule_root: next[rule_root]})
-                pending_facts.update(next[rule_root].keys())
-            else:
-                implicants, antecedents = next["__result__"]
-                if self.print_steps:
-                    pred_implicants = [self._to_pred(imp) for imp in implicants]
-                    pred_antecedents = [self._to_pred(ant) for ant in antecedents]
-                    sources = set(source for ant in antecedents for source in self.knowledge_base[ant])
-                    for imp in pred_implicants:
-                        print(f"Derived {imp} from {pred_antecedents} which was sourced from {sources}")
-
-
-        source_facts = set.union(*[self.knowledge_base[ant] for ant in antecedents], set())
-
-        res = (True, None)
-        for fact in implicants:
-            if -fact in self.knowledge_base:
-                if self.print_steps:
-                    explanation = self.knowledge_base[trig_fact] | self.knowledge_base[-fact]
-                    print(f"Found conflict triggered by {self._to_pred(trig_fact)} from {self._to_pred(fact)} caused by {explanation}")
-                    #print(f"Found conflict caused by {explanation}")
-                explanation = self.knowledge_base[trig_fact] | self.knowledge_base[-fact]
-                # res = self._assert_lit(expr, fact, source_facts)
-                #assert len(explanation) >= 2
-                res = (False, [-lit for lit in explanation])
-                break
-            pending_facts.add(fact)
-            self.add_fact(fact, source_facts)
-
-        return res, source_facts, pending_facts
-
-    # def traverse_rule_tree(self, rule_tree, current_facts):
-    #     """Recursively traverse the rule tree to check for matching conditions."""
-    #     if "__result__" in rule_tree:  # Consequence found
-    #         return rule_tree["__result__"]
-    #     for key, sub_tree in rule_tree.items():
-    #         if key in current_facts:  # Move deeper if condition exists
-    #             result = self.traverse_rule_tree(sub_tree, current_facts)
-    #             if result:
-    #                 return result
-    #     return None
-
-    def infer_facts(self):
-        """Infer new facts based on existing knowledge."""
-        new_facts = set()
-
-
-        # Traverse from each known fact
-        for fact in self.knowledge_base:
-            result = self.traverse_rule_tree(self.rules.get(fact, {}), self.knowledge_base)
-            if result and result not in self.knowledge_base:
-                new_facts.add(result)
-
-        # Update knowledge base
-        self.knowledge_base.update(new_facts)
-        return new_facts
-
-    def run_inference_until_stable(self):
-        """Keep running inference until no new facts are inferred."""
-        while True:
-            new_facts = self.infer_facts()
-            if not new_facts:
-                break
-
-    def __repr__(self):
-        """Return a string representation of the current knowledge base."""
-        return f"Knowledge Base: {self.knowledge_base}"
-
-rules_engine = RulesEngine(id_rules_dict)
+rules_engine = RulesEngine(id_rules_dict, direct_dict)
 
 class FCSolver():
     """
